gather: moved mDNS discovery to a separate file.
